+++ packages/bluesky/lib/src/services/codegen/app/bsky/feed/defs/viewer_state.dart
@@ -12,6 +12,9 @@ import 'package:atproto_core/atproto_core.dart';
 import 'package:atproto_core/internals.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+// Project imports:
+import './known_likers.dart';
+
 part 'viewer_state.freezed.dart';
 part 'viewer_state.g.dart';
 
@@ -30,6 +33,7 @@ abstract class ViewerState with _$ViewerState {
     'replyDisabled',
     'embeddingDisabled',
     'pinned',
+    'knownLikers',
   ];
 
   @JsonSerializable(includeIfNull: false)
@@ -42,6 +46,9 @@ abstract class ViewerState with _$ViewerState {
     bool? replyDisabled,
     bool? embeddingDisabled,
     bool? pinned,
+
+    /// This property is present only in selected cases, as an optimization.
+    @KnownLikersConverter() KnownLikers? knownLikers,
 
     Map<String, dynamic>? $unknown,
   }) = _ViewerState;
@@ -70,6 +77,8 @@ extension ViewerStateExtension on ViewerState {
   bool get isNotEmbeddingDisabled => !isEmbeddingDisabled;
   bool get isPinned => pinned ?? false;
   bool get isNotPinned => !isPinned;
+  bool get hasKnownLikers => knownLikers != null;
+  bool get hasNotKnownLikers => !hasKnownLikers;
 }
 
 final class ViewerStateConverter